% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{append.trees}
\alias{append.trees}
\title{Append isolation trees from one model into another}
\usage{
append.trees(model, other)
}
\arguments{
\item{model}{An Isolation Forest model (as returned by function \link{isolation.forest})
to which trees from `other` (another Isolation Forest model) will be appended into.}

\item{other}{Another Isolation Forest model, from which trees will be appended into
`model`.}
}
\value{
The updated `model` object, to which `model` needs to be reassigned
(i.e. you need to use it as follows: `model <- append.trees(model, other)`).
}
\description{
This function is intended for merging models \bold{that use the same hyperparameters} but
were fitted to different subsets of data.
In order for this work, both models must have been fit to data in the same format - 
that is, same number of columns, same order of the columns, and same types. This will
only work for numeric types - if the input data contains categorical columns, this is
likely to fail as the categories are internally re-indexed, and will give wrong results
unless both have the exact same factors appearing in the same order in the data (you can
check this through the results from `factor(data$col)`, which this library uses internally).
Using `factor` types with the same levels will \bold{not} make it work.

Note that this function will not perform any checks on the inputs, and passing two incompatible
models (e.g. fit to different numbers of columns) will result in wrong results and
potentially crashing the R process when using it.

\bold{Important:} the result of this function must be reassigned to `model` in order for it
to work properly - e.g. `model <- append.trees(model, other)`.
}
\examples{
library(isotree)

### Generate two random sets of data
m <- 100
n <- 2
set.seed(1)
X1 <- matrix(rnorm(m*n), nrow=m)
X2 <- matrix(rnorm(m*n), nrow=m)

### Fit a model to each dataset
iso1 <- isolation.forest(X1, ntrees=3, nthreads=1)
iso2 <- isolation.forest(X2, ntrees=2, nthreads=1)

### Check the terminal nodes for some observations
nodes1 <- predict(iso1, head(X1, 3), type="tree_num")
nodes2 <- predict(iso2, head(X1, 3), type="tree_num")

### Append the trees from 'iso1' into 'iso1'
iso1 <- append.trees(iso1, iso2)

### Check that it predicts the same as the two models
nodes.comb <- predict(iso1, head(X1, 3), type="tree_num")
nodes.comb$tree_num == cbind(nodes1$tree_num, nodes2$tree_num)

### The new predicted scores will be a weighted average
### (Be aware that, due to round-off, it will not match with '==')
nodes.comb$score
(3*nodes1$score + 2*nodes2$score) / 5
}
