% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{deepcopy.isotree}
\alias{deepcopy.isotree}
\title{Deep-Copy an Isolation Forest Model Object}
\usage{
deepcopy.isotree(model)
}
\arguments{
\item{model}{An `isolation_forest` model object.}
}
\value{
A new `isolation_forest` object, with deep-copied C++ objects.
}
\description{
Deep-Copy an Isolation Forest Model Object
}
\details{
Generates a deep copy of a model object, including the C++ objects inside it.
This function is only meaningful if one intends to call a function that modifies the
internal C++ objects - currently, the only such function is \link{add.isolation.tree} - as
otherwise R's objects follow a copy-on-write logic.
}
