% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{get.num.nodes}
\alias{get.num.nodes}
\title{Get Number of Nodes per Tree}
\usage{
get.num.nodes(model)
}
\arguments{
\item{model}{An Isolation Forest model as produced by function `isolation.forest`.}
}
\value{
A list with entries `"total"` and `"terminal"`, both of which are integer vectors
with length equal to the number of trees. `"total"` contains the total number of nodes that
each tree has, while `"terminal"` contains the number of terminal nodes per tree.
}
\description{
Get Number of Nodes per Tree
}
