% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.deep.copy}
\alias{isotree.deep.copy}
\title{Deep-Copy an Isolation Forest Model Object}
\usage{
isotree.deep.copy(model)
}
\arguments{
\item{model}{An `isolation_forest` model object.}
}
\value{
A new `isolation_forest` object, with deep-copied C++ objects.
}
\description{
Generates a deep copy of a model object, including the C++ objects inside it.
This function is only meaningful if one intends to call a function that modifies the
internal C++ objects - currently, the only such function are \link{isotree.add.tree}
and \link{isotree.append.trees} - as otherwise R's objects follow a copy-on-write logic.
}
