% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.drop.reference.points}
\alias{isotree.drop.reference.points}
\title{Drop Reference Points from Isolation Forest Model Object}
\usage{
isotree.drop.reference.points(model)
}
\arguments{
\item{model}{An `isolation_forest` model object.}
}
\value{
The same `model` object, but now with the reference points removed. \bold{Note that `model` is modified in-place
in any event}.
}
\description{
Drops any reference points used for distance and/or kernel calculations
from the model object, if any were set through \link{isotree.set.reference.points}.
}
\seealso{
\link{isotree.set.reference.points}
}
