% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.set.nthreads}
\alias{isotree.set.nthreads}
\title{Set Number of Threads for Isolation Forest Model Object}
\usage{
isotree.set.nthreads(model, nthreads = 1L)
}
\arguments{
\item{model}{An Isolation Forest model (as returned by function \link{isolation.forest})
for which an indexer for terminal node numbers and/or distances will be added.

\bold{The object will be modified in-place}.}

\item{nthreads}{Number of threads to set for this model object to use.}
}
\value{
The same `model` object (as invisible), but now with a different configured number of threadst. Note
the input object is modified in-place regardless.
}
\description{
Changes the number of threads that an isolation forest model object
will use when calling functions such as `predict`.
}
