% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.subset.trees}
\alias{isotree.subset.trees}
\alias{[.isolation_forest}
\title{Subset trees of a given model}
\usage{
isotree.subset.trees(model, trees_take)

\method{[}{isolation_forest}(x, i)
}
\arguments{
\item{model, x}{An `isolation_forest` model object.}

\item{trees_take, i}{Indices of the trees of `model` to copy over to a new model,
as an integer vector.
Must be integers with numeration starting at one}
}
\value{
A new isolation forest model object, containing only the subset of trees
from this `model` that was specified under `trees_take`.
}
\description{
Creates a new isolation forest model containing only selected trees of a
given isolation forest model object. Note that, if using `lazy_serialization=FALSE`,
this will re-trigger serialization.
}
