% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{length.isolation_forest}
\alias{length.isolation_forest}
\title{Get Number of Trees in Model}
\usage{
\method{length}{isolation_forest}(x)
}
\arguments{
\item{x}{An isolation forest model, as returned by function \link{isolation.forest}.}
}
\value{
The number of trees in the model, as an integer.
}
\description{
Returns the number of trees in an isolation forest model.
}
