% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{summary.isolation_forest}
\alias{summary.isolation_forest}
\title{Print summary information from Isolation Forest model}
\usage{
\method{summary}{isolation_forest}(object, ...)
}
\arguments{
\item{object}{An Isolation Forest model as produced by function `isolation.forest`.}

\item{...}{Not used.}
}
\value{
No return value.
}
\description{
Displays the most general characteristics of an isolation forest model (same as `print`).
}
\details{
Note that after loading a serialized object from `isolation.forest` through `readRDS` or `load`,
it will only de-serialize the underlying C++ object upon running `predict`, `print`, or `summary`,
so the first run will be slower, while subsequent runs will be faster as the C++ object will already be in-memory.
}
\seealso{
\link{isolation.forest}
}
