% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{variable.names.isolation_forest}
\alias{variable.names.isolation_forest}
\title{Get Variable Names for Isolation Forest Model}
\usage{
\method{variable.names}{isolation_forest}(object, ...)
}
\arguments{
\item{object}{An isolation forest model, as returned by function \link{isolation.forest}.}

\item{...}{Not used.}
}
\value{
A character vector containing the column / variable names.
}
\description{
Returns the names of the input data columns / variables to which an
isolation forest model was fitted.

If the data did not have column names, it will make them up as "column_1..N".

Note that columns will always be reordered so that numeric columns come first, followed by
categorical columns.
}
