% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.drop.imputer}
\alias{isotree.drop.imputer}
\title{Drop Imputer Sub-Object from Isolation Forest Model Object}
\usage{
isotree.drop.imputer(model)
}
\arguments{
\item{model}{An `isolation_forest` model object.}
}
\value{
The same `model` object, but now with the imputer removed. \bold{Note that `model` is modified in-place
in any event}.
}
\description{
Drop Imputer Sub-Object from Isolation Forest Model Object
}
\details{
Drops the imputer sub-object from an isolation forest model object, if it was fitted with data imputation
capabilities. The imputer, if constructed, is likely to be a very heavy object which might
not be needed for all purposes.
}
