% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.drop.indexer}
\alias{isotree.drop.indexer}
\title{Drop Indexer Sub-Object from Isolation Forest Model Object}
\usage{
isotree.drop.indexer(model)
}
\arguments{
\item{model}{An `isolation_forest` model object.}
}
\value{
The same `model` object, but now with the indexer removed. \bold{Note that `model` is modified in-place
in any event}.
}
\description{
Drops the indexer sub-object from an isolation forest model object, if it was constructed.
The indexer, if constructed, is likely to be a very heavy object which might
not be needed for all purposes.
}
\details{
Note that reference points as added through \link{isotree.set.reference.points} are
associated with the indexer object and will also be dropped if any were added.
}
\seealso{
\link{isotree.build.indexer}
}
