% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isoforest.R
\name{isotree.subset.trees}
\alias{isotree.subset.trees}
\title{Subset trees of a given model}
\usage{
isotree.subset.trees(model, trees_take)
}
\arguments{
\item{model}{An `isolation_forest` model object.}

\item{trees_take}{Indices of the trees of `model` to copy over to a new model,
as an integer vector.
Must be integers with numeration starting at one}
}
\value{
A new isolation forest model object, containing only the subset of trees
from this `model` that was specified under `trees_take`.
}
\description{
Subset trees of a given model
}
\details{
Creates a new isolation forest model containing only selected trees of a
given isolation forest model object.
}
